Remove need for `GITHUB_TOKEN` for private repositories.
Use the fact that `WORKDIR` of custom Docker Github Action steps has the
`GITHUB_WORKDIR` mounted to it, and that `truffleHog` allows `file://`
URLs to be used to workaround needing a `GITHUB_TOKEN` that has clone
perms.

This should make setting up this action, for public and private repos,
the same and much easier.